Feat server.mjs is entrypoint file
